ga_rpc: initial ga_rpc class
This will be our interface to gdk_rpc calls. rpc technically session-less, but
perhaps it will be good to treat it like a session, to maintain connectivity
with heartbeats.

This can probably be handled on the gdk_rpc side, but let's just call it a
session for now and treat it that way.